Replaced Chewfile with Insheet
Replace chewfile with insheet to avoid problems when insheet imports
variables not as strings.

Also made it so every column gets it own alignment parameter so that
tables look better.  A bit messy in the output, but the whole point is
making this automatically beautiful.

3 - added package {caption} to align captions to center even if they are
over two lines or more.