Add files via upload
This adds the option "package(packagelist)" to tex3pt. This option allows users to specify packages to include in the preamble.
An example for the use of the option are tables with checkmarks. For checkmarks to be displayed the user may want to include the \usepackage{"amssymb"}. This would be done with:
tex3pt table using filename , package(amssymb) preamb

It may be useful to change this functionality to allow the user to add more flexible input to the preamble.